TizenRefApp-6625 Newly created contact isn't displayed in search results

+++ lib-common/src/Contacts/Model/SearchEngine.cpp
@@ -207,7 +207,7 @@ void SearchEngine::historyFor(HistoryForFn function)
 		auto &list = m_History[i];
 		query.append(1, m_Query[i]);
 
-		if (i < (size_t)m_LastFoundIndex && list.empty()) {
+		if ((int) i < m_LastFoundIndex && list.empty()) {
 			continue;
 		}
